Updated scrollable.tcl
Scrollable has been updated so that it binds the resizemejig to the configure event of the frame or the canvas.

This means that the frame is now always as wide as the canvas, and the scrollarea will always be as tall as it needs to be.

A "commented" out section has been left in.
It does for height what has been done for width.

It can be used, but in this case it seems to work better without it.
Maybe.