Made more complex package description.
I've replaced the Package Description label with a frame.
This frame has in it fields that package description data should go in.

Also, I've changed the method by which those labels are set.
Originally, they were being set by calling the configure method on the label.

Now, I'm using the textvariable property of the label.

Side-Effects abound.
Basically, the reason is two-fold.
First, it's a global value, so you don't have to worry about the address of the box of whatever.
You just set the variable and it will worry about displaying it.

Second, it's far easier to attach to a variable than a box configure command.
If later you wanted to set another label somewhere and have it remain in sync, you just attach it to the same value.
If you want to set the window's title to whatever the current package is, you can put a trace of the currentPackage(title) variable, and set the title whenever it's changed.
Stuff like that.