Connect the category box to an actual filter.
I've made a lot of changes to how the combo box works so that it can work.
I've got a map from a display value to a data value, so we can show a more friendly value to the user.

Also, when "All" is clicked in the category, it shows the text "Category".
This will allow me to remove the label later.