New utility to get package dependencies.
Pass it the name of a package, and it will list the dependencies that
need to be installed for that package.

External dependencies will be output as:
E PACKAGE VERSION

Internal dependencies which are already up to date will not be output.

Internal dependencies which need to be installed/updated will output as:
I PACKAGE VERSION

Each dependency will only be output once.

A package is not it's own dependency.

Virtual packages are handled invisibly.  Output package names match the
"real" package name.