Added new gui helper function file.
I'm about to make the long description use a textbox instead of some hacked label.

The main problem with a textbox was the difficulty of putting values into it.

I've fixed that with this file.
It replicates the '-textvariable' option's functionality in labels for textboxes.

It sets up a trace on a variable, and anytime that variable's value changes, the textbox is updated to show its value.

It only goes one way, it is only for displaying a value in a textbox.