Better Handling of Search Backend Command Construction.
I realized during the last task that the search wasn't working anymore.
I don't know why that was, but I fixed what I felt to be too raw anyway.
It works now.

In this version, instead of putting crap into variables and then putting those into a string, we build up a list and then execute that.
It seems much cleaner and more foreward compatible.

It seems to be working so far.