Diff - Backend
At this point it just spits out a list of packages to stdout with the package name and its requested status.

0 for uninstall, 1 for install.

It's not the most stable system... but it's fine for now.
It can be screwed up if some variable changes without the checkbutton being clicked.

If this is a problem, it will need to be stabelized later.