Fix a bug where the package list can be scrolled entirely off screen.
This mostly happens when you narrow a search while at the bottom of the list.
The problem was that it was drawing the list before all of the frames has a chance to compute their sizes.

Now, we wait 100ms before resizing everything, so the canvas takes care of all the "off-the-edge" stuff.