Actual Ability to Install and Remove.
Currently only works if you already have the ability to write to the TVEROOT directory and all sub directories etc.
That is, it does no "sudo" like functionality.

Also, there seems to be a problem with my tracking of changes.
I think I've done something wrong in my reseting of statuses.
This should be fixed in a later change.