Added Logout Button.
I also added a logout script to handle the logging out.

Also, this allows one to go from offline to online.
This isn't really a logout at all.
Instead it checks if you're logged in now, perhaps through some other means or change in network state, and if so draws the UI.
If you're still not logged in, it just draws the login screen.
From there you can attempt to log in, or you can go back to offline mode.