Wrote the Algorithm to Draw Stars.
I was going to use images previously, but in the end I think this is the most flexible.
It allows one to, on the fly, change settings about colour or size without having to deal with images.

Right now all of the configuration options are specified in the function with no way to change them from the outside.
This is intended to change.