Made The Login Check More Asynchronous, and Skippable.
It seemed like the UI was failing to open, but then I realized it's because my crappy internet connection was making login-check block.

So, if I waited until the connection timed-out it would work.

I said to myself "This will not do".

So, I asynchronized the login check and made a window that informs the user what's going on.
It also gives them the option to skip the check and just go offline.

This can be used by people who either don't have the internet, or know they're on an unreliable connection.