Made the Rating Widget's ReadOnly Option Configurable After Creation.
It turns out I was a little Too efficient.

I decided it might be nice if I could change whether or not the widget is readonly after it exists.
The way I was doing it, though, was to just not make the bindings if it was readonly.
This isn't very reconfigurable, though.

So, the bindings get made no matter what, now, and they just don't do anything if the widget is readonly.

Then, the 'configuration' command is intercepted and it's checked to see if the thing you're configuring is -readonly.
If so, it's changed.
If not, the whole thing is passed along to the standard configuration command.