Updated TODO.
First off, I completed the Rating UI.

Also, it's come to my attention that the User's ratings are, in fact, fetched dynamically.
The TODO has been updated to reflect this.

The Feedback comment has been reformatted to be more in-keeping with the rest of them.

Also, I relaized that I don't have any UI in place for handling updates.
Now it's a TODO.