Made calculateTotal Use Purchased Status rather than Installed Status.
This was not a good way to do it, but there wasn't a better way at the time.
Now there is!

For now it assumes that if you own one version you own them all.
This is valid until we do upgrades that cost, then this will need to be updated.