Fixed Outputting of Purchase.
I was using '-f' to detect whether or not curl 'failed'.
That is, if the status was other than 200.

The problem, though, is that when it failed it returned nothing.
That wasn't great, because then I didn't know why it failed.

Now I've changed it so I get the status code and the reponse.
Now I can compare the reponse myself.