Better Representation of Status.
Shows purchase.png when the person owns the package and update.png when there's an update available.

I considered for a bit whether or not to add the png files.
Technically they're the output, with the SVG being the source.
I figured they wouldn't change a lot though, and it's not strictly easy to do the conversion without booting up InkScape.
So, I added the pngs and the svgs.

If that's a problem, it may be rectified. I'll understand.