Made the Tcl Includes a Package.
Kind of a small hack around the package system.
I've made a "Package" called tveLibs which just has all the includes.

So, it looks to see if there's a directory in argv0 that is called tveTclIncludes.
If there is, it adds that dir to the package include path.

In either case, it tries to load the package.
This way, it can be installed anywhere that's in auto_path on the host system.
(I think only one dir deep)